+++ sensu/source/CHANGELOG.md
@@ -3,6 +3,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
+## [0.5.0] - 2018-07-16
+- Made debug flag boolean
+- Added additional debug logging for troubleshooting
+- Added Unit Tests
+
 ## [0.4.0] - 2018-07-13
 - Added handling for client environments set as array of values
 - Added schemas so graphite metric names can be parsed into dimensional metrics